apps/python/requirements.txt
primate-run~=0.6.1
